+++ homeworks/emiliano-acerbi/20-routing-redux/redux/src/App.jsx
@@ -1,50 +1,23 @@
 import React, { useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { addTodo } from './features/todo/todoSlice';
-import { v4 as uuidv4 } from 'uuid';
+import { useSelector } from 'react-redux';
+import Container from './components/Container';
 import Todo from './components/Todo';
+import TodoForm from './components/TodoForm';
 
 function App() {
   const [inputValue, setInputValue] = useState('');
 
   const todos = useSelector((state) => state.todos);
-  const dispatch = useDispatch();
-
-  const handleAddTodo = (e) => {
-    e.preventDefault();
-    const newTodo = {
-      id: uuidv4(),
-      text: inputValue,
-    };
-
-    dispatch(addTodo(newTodo));
-    setInputValue('');
-  };
 
   return (
-    <main className="min-h-screen bg-gradient font-primary text-3xl flex flex-col items-center gap-5">
-      <form className="flex mt-28">
-        <input
-          className="w-[554px] h-[65px] px-5 outline-none"
-          placeholder="Create Todo-Task"
-          value={inputValue}
-          onChange={(e) => setInputValue(e.target.value)}
-        />
-        <button
-          className="bg-add-bg hover:opacity-80 text-white w-[122px] h-[65px]"
-          type="submit"
-          onClick={handleAddTodo}
-        >
-          Add
-        </button>
-      </form>
-
+    <Container>
+      <TodoForm inputValue={inputValue} setInputValue={setInputValue} />
       <div className="flex flex-col gap-10 mt-20">
         {todos.map((todo) => (
           <Todo key={todo.id} todo={todo} inputValue={inputValue} />
         ))}
       </div>
-    </main>
+    </Container>
   );
 }
